% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid3Covariates.R
\name{grid3ListCountryCovariates}
\alias{grid3ListCountryCovariates}
\title{grid3ListCountryCovariates function will return a data frame of
avalible covariates for a country}
\usage{
grid3ListCountryCovariates(ISO3 = NULL, username = NULL, password = NULL,
  detailed = FALSE, frCSVDownload = FALSE)
}
\arguments{
\item{ISO3}{a 3-character country code or vector of country codes}

\item{username}{ftp username to WorldPop ftp server}

\item{password}{ftp password to WorldPop ftp server}

\item{detailed}{If TRUE, then more information will be given}

\item{frCSVDownload}{If TRUE, a new wpgAllCovariates.csv file will downloaded}
}
\value{
Dataframe
}
\description{
grid3ListCountryCovariates function will return a data frame of
avalible covariates for a country
}
\examples{
grid3ListCountryCovariates( ISO3="USA", username="ftpUsername", password="ftpPassword" )

grid3ListCountryCovariates(ISO3=c("USA","AFG"), username="ftpUsername", password="ftpPassword" )
}
